+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 * Added `FullVersion::parse` convenience method
 * Added `BaseVersion::parse` convenience method
+* Added `Version::map` method
+* Added `BaseVersion::map` method
+* Added `FullVersion::map` method
 
 ## [0.3.0]